+++ src/Components/search/input/searchInput.js
@@ -1,10 +1,5 @@
 const searchTemplate = document.querySelector("template");
-const searchData = {
-    array: "",
-    input: "",
-    container: "",
-    template: "",
-}
+const noResults = document.querySelector(".noResults");
 
 function searchBar() {
     const searchBar = searchTemplate.content.cloneNode(true);
@@ -15,13 +10,23 @@ function searchBar() {
 
 searchBar();
 
+const searchInput = document.querySelector(".search-input");
+const body = document.body;
+
+const searchData = {
+    array: "",
+    input: searchInput,
+    container: body,
+    template: noResults,
+}
+
 function search({ array, input, container, template }) {
     let searchedVal = input.value.trim().toLowerCase();
     const searched = array.filter(arr => {
-        return arr.value.toLowerCase().includes(searchedVal);
+        return arr.name.toLowerCase().startsWith(searchedVal);
     });
 
-    render(container, searched);
+    render(body, searched);
 
     if (searched.length === 0) {
         container.innerHTML = "";